python/requirements.txt
# HTTP Client for Laravel API communication
requests==2.31.0
# httpx==0.27.0  # Commented out - causes asyncio event loop conflicts with Playwright Sync API

# Playwright for browser automation
playwright==1.48.0

# Environment variables
python-dotenv==1.0.1

# Database (optional, if direct DB access needed)
# mysql-connector-python==8.3.0
# pymysql==1.1.1

# LLM API clients
# openai==1.51.0  # Commented out - not used and may create asyncio event loops

# Utilities
beautifulsoup4==4.12.3
lxml==5.3.0

# Logging
loguru==0.7.2

# Retry logic
tenacity==9.0.0

# Machine Learning
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3

# Advanced ML Models (optional, will fallback if not available)
xgboost==2.0.3
lightgbm==4.1.0

# Visualization (optional)
matplotlib==3.7.2
seaborn==0.12.2

